Former-commit-id: 827ba441bddcdcb3a8e3f7ee0fc2291427cbdfe6

+++ test/JDBCPrepareStatement.java
@@ -1454,7 +1454,7 @@ public class JDBCPrepareStatement {
 		if (pstmt.execute()) {
 			rs=pstmt.getResultSet();
 			while (rs.next()) {
-				org.junit.Assert.assertEquals(10001,rs.getInt(1));
+				org.junit.Assert.assertEquals(10001,rs.getLong(1));
 			}
 		}
 		if (rs != null) {
